infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-aws-s3>=0.28.0",
    "aws-cdk-lib>=2.183.0",
    "constructs>=10.4.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ty>=0.0.1a14",
]

[tool.uv]
managed = true
